feat: interface揃えるまで
