+++ Unity/Assets/AnKuchen/Scripts/KuchenList/IReusableMappedObject.cs
@@ -1,0 +1,10 @@
+using AnKuchen.Map;
+
+namespace AnKuchen.KuchenList
+{
+    public interface IReusableMappedObject : IMappedObject
+    {
+        void Activate();
+        void Deactivate();
+    }
+}
