+++ Unity/Assets/AnKuchen/Scripts/KuchenList/KuchenList.cs
@@ -1,0 +1,19 @@
+using UnityEngine;
+
+namespace AnKuchen.KuchenList
+{
+    public class KuchenList : MonoBehaviour
+    {
+        public IKuchenList List { get; set; }
+
+        public void OnDestroy()
+        {
+            List.OnDestroy();
+        }
+    }
+
+    public interface IKuchenList
+    {
+        void OnDestroy();
+    }
+}
