+++ Unity/Assets/AnKuchen/Scripts/KuchenList/Margin.cs
@@ -1,0 +1,10 @@
+namespace AnKuchen.KuchenList
+{
+    public class Margin
+    {
+        public float Top { get; set; }
+        public float Bottom { get; set; }
+        public float Left { get; set; }
+        public float Right { get; set; }
+    }
+}
