+++ Unity/Assets/AnKuchen/Scripts/KuchenList/VerticalList.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: e1443937cd6146ae8aa96d1ab3234918
+timeCreated: 1607498189
