+++ Unity/Assets/Development/Sample.cs
@@ -1,11 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Collections.Generic;
 using AnKuchen.Map;
 using UnityEngine;
 using UnityEngine.UI;
-using AnKuchen.Extensions;
-using Object = UnityEngine.Object;
+using AnKuchen.KuchenList;
 
 namespace AnKuchen.Development
 {
@@ -113,10 +110,12 @@ namespace AnKuchen.Development
 
         public void Activate()
         {
+            Debug.Log("Activate");
         }
 
         public void Deactivate()
         {
+            Debug.Log("Deactivate");
         }
     }
 
@@ -143,130 +142,5 @@ namespace AnKuchen.Development
         {
             Button.onClick.RemoveAllListeners();
         }
-    }
-}
-
-
-
-namespace AnKuchen
-{
-    public interface IReusableMappedObject : IMappedObject
-    {
-        void Activate();
-        void Deactivate();
-    }
-
-    public class UIFactory<T1, T2> where T1 : IMappedObject where T2 : IMappedObject
-    {
-        public Action<T1> Callback1 { get; }
-        public Action<T2> Callback2 { get; }
-
-        public UIFactory(Action<T1> callback1)
-        {
-            Callback1 = callback1;
-        }
-
-        public UIFactory(Action<T2> callback2)
-        {
-            Callback2 = callback2;
-        }
-    }
-
-    public class VerticalList<T1, T2> : IDisposable where T1 : IReusableMappedObject, new() where T2 : IReusableMappedObject, new()
-    {
-        private readonly ScrollRect scrollRect;
-        private readonly T1 original1;
-        private readonly T2 original2;
-        private List<UIFactory<T1, T2>> contents = new List<UIFactory<T1, T2>>();
-        private readonly List<(GameObject, IReusableMappedObject)> items = new List<(GameObject, IReusableMappedObject)>();
-
-        public VerticalList(ScrollRect scrollRect, T1 original1, T2 original2)
-        {
-            this.scrollRect = scrollRect;
-            this.original1 = original1;
-            this.original2 = original2;
-
-            this.original1.Mapper.Get().SetActive(false);
-            this.original2.Mapper.Get().SetActive(false);
-        }
-
-        public T1 Get1(int index)
-        {
-            return (T1) items[index].Item2;
-        }
-
-        public T2 Get2(int index)
-        {
-            return (T2) items[index].Item2;
-        }
-
-        private void ClearAll()
-        {
-            foreach (var item in items)
-            {
-                item.Item2.Deactivate();
-                Object.Destroy(item.Item1);
-            }
-
-            items.Clear();
-        }
-
-        private void UpdateListContents()
-        {
-            ClearAll();
-
-            foreach (var content in contents)
-            {
-                if (content.Callback1 != null) CreateNewObject(original1, content.Callback1);
-                if (content.Callback2 != null) CreateNewObject(original2, content.Callback2);
-            }
-        }
-
-        private void CreateNewObject<T>(T original, Action<T> contentCallback) where T : IReusableMappedObject, new()
-        {
-            var newObject = original.Duplicate();
-            newObject.Mapper.Get().SetActive(true);
-            newObject.Activate();
-            contentCallback(newObject);
-            items.Add((newObject.Mapper.Get(), newObject));
-        }
-
-        public ListContentEditor Edit()
-        {
-            return new ListContentEditor(this);
-        }
-
-        public class ListContentEditor : IDisposable
-        {
-            private readonly VerticalList<T1, T2> parent;
-            public List<UIFactory<T1, T2>> Contents { get; set; }
-            public float Spacing { get; set; }
-            public Margin Margin { get; set; } = new Margin();
-
-            public ListContentEditor(VerticalList<T1, T2> parent)
-            {
-                this.parent = parent;
-                Contents = parent.contents;
-            }
-
-            public void Dispose()
-            {
-                parent.contents = Contents;
-                parent.UpdateListContents();
-            }
-        }
-
-        public void Dispose()
-        {
-            ClearAll();
-        }
-    }
-
-    public class Margin
-    {
-        public float Top { get; set; }
-        public float Bottom { get; set; }
-        public float Left { get; set; }
-        public float Right { get; set; }
     }
 }
