+++ Unity/Assets/Development/SampleScene.unity
@@ -427,9 +427,9 @@ RectTransform:
   m_Father: {fileID: 1771712974}
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
-  m_AnchoredPosition: {x: 0, y: 0}
+  m_AnchorMin: {x: 0.5, y: 0.5}
+  m_AnchorMax: {x: 0.5, y: 0.5}
+  m_AnchoredPosition: {x: -50, y: 400}
   m_SizeDelta: {x: 100, y: 100}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!1 &258812533
@@ -506,12 +506,18 @@ MonoBehaviour:
     Path: 1ce69cde010a3f2f7d268c36
   - GameObject: {fileID: 1640439751}
     Path: 8cad3b581ce69cde010a3f2f7d268c36
+  - GameObject: {fileID: 443385791}
+    Path: d445594f8cad3b581ce69cde010a3f2f7d268c36
   - GameObject: {fileID: 2029483824}
     Path: 266ab3018cad3b581ce69cde010a3f2f7d268c36
   - GameObject: {fileID: 245399791}
     Path: 3d2773f61ce69cde010a3f2f7d268c36
   - GameObject: {fileID: 1813606599}
     Path: d445594f3d2773f61ce69cde010a3f2f7d268c36
+  - GameObject: {fileID: 93207777}
+    Path: 3d68bf5a3d2773f61ce69cde010a3f2f7d268c36
+  - GameObject: {fileID: 943476767}
+    Path: 266ab3013d68bf5a3d2773f61ce69cde010a3f2f7d268c36
   - GameObject: {fileID: 674873205}
     Path: 1fa7559c7d268c36
   - GameObject: {fileID: 1629945293}
@@ -994,8 +1000,8 @@ MonoBehaviour:
   m_TargetGraphic: {fileID: 1652613390}
   m_HandleRect: {fileID: 1652613389}
   m_Direction: 2
-  m_Value: 0
-  m_Size: 1
+  m_Value: 1.0000001
+  m_Size: 0.3987317
   m_NumberOfSteps: 0
   m_OnValueChanged:
     m_PersistentCalls:
@@ -1530,9 +1536,9 @@ RectTransform:
   m_Father: {fileID: 1771712974}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
-  m_AnchoredPosition: {x: 0, y: 0}
+  m_AnchorMin: {x: 0.5, y: 0.5}
+  m_AnchorMax: {x: 0.5, y: 0.5}
+  m_AnchoredPosition: {x: -50, y: 475}
   m_SizeDelta: {x: 100, y: 50}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!1 &1652613388
@@ -1740,8 +1746,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 1771712974}
-  - component: {fileID: 1771712976}
-  - component: {fileID: 1771712975}
   m_Layer: 5
   m_Name: Content
   m_TagString: Untagged
@@ -1765,50 +1769,11 @@ RectTransform:
   m_Father: {fileID: 169310655}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0, y: 1}
-  m_AnchorMax: {x: 1, y: 1}
-  m_AnchoredPosition: {x: 0, y: 0}
-  m_SizeDelta: {x: 0, y: 0}
+  m_AnchorMin: {x: 0.5, y: 0.5}
+  m_AnchorMax: {x: 0.5, y: 0.5}
+  m_AnchoredPosition: {x: -92, y: 199.99995}
+  m_SizeDelta: {x: 200, y: 1000}
   m_Pivot: {x: 0, y: 1}
---- !u!114 &1771712975
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1771712973}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 59f8146938fff824cb5fd77236b75775, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  m_Padding:
-    m_Left: 0
-    m_Right: 0
-    m_Top: 0
-    m_Bottom: 0
-  m_ChildAlignment: 0
-  m_Spacing: 0
-  m_ChildForceExpandWidth: 1
-  m_ChildForceExpandHeight: 1
-  m_ChildControlWidth: 0
-  m_ChildControlHeight: 0
-  m_ChildScaleWidth: 0
-  m_ChildScaleHeight: 0
---- !u!114 &1771712976
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1771712973}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3245ec927659c4140ac4f8d17403cc18, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  m_HorizontalFit: 2
-  m_VerticalFit: 2
 --- !u!1 &1813606599
 GameObject:
   m_ObjectHideFlags: 0
